added gdisconnect

+++ finalProject.py
@@ -114,6 +114,40 @@ def gconnect():
     print "done!"
     return output
 
+# Revoke the user's token and reset login session
+@app.route('/gdisconnect')
+def gdisconnect():
+        # Only disconnect a connected user.
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = credentials.access_token
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+
+    if result['status'] == '200':
+        # Reset the user's sesson.
+        del login_session['credentials']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+        # For whatever reason, the given token was invalid.
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+
 # XML for the entire menu
 @app.route('/menu/XML')
 def menuXML():
@@ -144,6 +178,9 @@ def menu():
 @app.route('/menu/<string:category>/new/',
         methods=['GET', 'POST'])
 def newMenuItem(category):
+    #  Prompt the user to login
+    if 'username' not in login_session:
+        return redirect('/login')
     # Add new menu item to this category type
     if request.method == 'POST':
         newItem = MenuItem(name=request.form['name'],
@@ -161,6 +198,9 @@ def newMenuItem(category):
 @app.route('/menu/<string:category>/<int:item_id>/edit/',
         methods=['GET', 'POST'])
 def editMenuItem(category, item_id):
+    #  Prompt the user to login
+    if 'username' not in login_session:
+        return redirect('/login')
     # Edit menu item in this category type
     editedItem = session.query(MenuItem).filter_by(id=item_id).one()
     if request.method == 'POST':
@@ -183,6 +223,9 @@ def editMenuItem(category, item_id):
 @app.route('/menu/<string:category>/<int:item_id>/delete/',
         methods=['GET', 'POST'])
 def deleteMenuItem(category, item_id):
+    #  Prompt the user to login
+    if 'username' not in login_session:
+        return redirect('/login')
     # Delete this menu item in this category type
     deleteItem = session.query(MenuItem).filter_by(id=item_id).one()
     if request.method == 'POST':
